requirements.txt
torch==2.8.0
numpy==1.26.4
scipy==1.13.1
PyYAML==6.0.2

